requirements.txt
openai
llama-index
trulens-eval
python-dotenv
pypdf
torch
torch
torchvision
torchaudio
numpy
pandas
sentence-transformers
sentencepiece
